+++ resources/views/unlayer.blade.php
@@ -3,19 +3,23 @@
 @endpush
 
 <script>
-    window.initialized = false;
+    window.unlayerInitialized = false;
 
     document.getElementById('unlayer').addEventListener('load', initUnlayer);
+
+    document.addEventListener('turbolinks:before-visit', confirmBeforeLeaveAndDestroyUnlayer);
     document.addEventListener("turbolinks:load", initUnlayer);
+    window.addEventListener('beforeunload', confirmBeforeLeaveAndDestroyUnlayer);
+
 
     function initUnlayer() {
-        if (window.initialized) {
+        if (window.unlayerInitialized) {
             return;
         }
 
         unlayer.init({!! json_encode($options) !!});
 
-        window.initialized = true;
+        window.unlayerInitialized = true;
 
         unlayer.loadDesign({!! $structuredHtml !!});
 
@@ -58,14 +62,33 @@
             unlayer.exportHtml(function(data) {
                 document.getElementById('html').value = data.html;
                 document.getElementById('structured_html').value = JSON.stringify(data.design);
+                document.getElementById('html').dataset.dirty = "";
                 document.querySelector('.layout-main form').submit();
             });
         });
+
+        unlayer.addEventListener('design:updated', function(data) {
+            document.getElementById('html').dataset.dirty = "dirty";
+        });
     }
+
+    function confirmBeforeLeaveAndDestroyUnlayer(event) {
+        if (document.getElementById('html').dataset.dirty === "dirty" && ! confirm('Are you sure you want to leave this page? Any unsaved changes will be lost.')) {
+            event.preventDefault();
+            return;
+        }
+
+        window.unlayerInitialized = false;
+
+        document.removeEventListener('turbolinks:before-visit', confirmBeforeLeaveAndDestroyUnlayer);
+        document.removeEventListener("turbolinks:load", initUnlayer);
+        window.removeEventListener('beforeunload', confirmBeforeLeaveAndDestroyUnlayer);
+    }
+
 </script>
 <div>
     <div class="form-row max-w-full h-full">
-        <label class="label" for="html">Body</label>
+        <label class="label" for="html">{{ __('Body') }}</label>
         @isset($errors)
             @error('html')
                 <p class="form-error" role="alert">{{ $message }}</p>
@@ -81,7 +104,7 @@
 
 <div class="form-buttons">
     <button id="save" type="submit" class="button">
-        <x-icon-label icon="fa-code" text="Save content"/>
+        <x-icon-label icon="fa-code" :text="__('Save content')"/>
     </button>
 </div>