+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to `laravel-mailcoach-unlayer` will be documented in this file
 
+## 1.1.2 - 2020-07-01
+
+- Confirm to leave when there are unsaved changes
+
+## 1.1.1 - 2020-06-12
+
+- fix Turbolinks compatibility
+
+## 1.1.0 - 2020-06-10
+
+- add translations (#1)
+
 ## 1.0.1 - 2020-03-18
 
 - make sure images keep their original format